Update templates.md
The "Extend from a template" entire section was not showing up in the document because it was missing a hash. 
And because this section was not showing up, the link in the document "extending from a template" under "Variable reuse" section is also broken. 

So enabling the "Extend from a template" by adding the additional #.